refactor: type 폴더명 types로 수정
