+++ src/features/cart/ui/CartItemQuantitySelector.tsx
@@ -1,5 +1,5 @@
 import * as S from './CartItemQuantitySelector.styles';
-import { CartItem } from '../type/cart';
+import { CartItem } from '../types/cart';
 import { useCartItemsContext } from '../context/useCartItemsContext';
 
 interface CartItemQuantitySelectorProps {
