+++ src/features/cart/utils/priceCalculations.ts
@@ -1,4 +1,4 @@
-import { CartItem } from '../type/cart';
+import { CartItem } from '../types/cart';
 
 export const calculatePrices = (selectedCartItems: CartItem[]) => {
   const cartTypeQuantity = selectedCartItems.length;
